feat!: add an NBT parser and allow to load .cw world files!
ClassicWorld is a world format used by various classic servers and
clients. it's pretty much just the byte block array packed into an
NBT with some extra metadata stuff.

considering that most modern clients support it, there's no point in
implementing any other format than this one (plus it's probably the
cleanest format to implement)

+++ Tako.Server/Game/World/WorldStreamer.cs
@@ -100,7 +100,7 @@ public class WorldStreamer
 		do
 		{
 			var amount = Math.Min(1024, size - cursor);
-			var percent = (byte)Math.Round(((cursor + amount) / size) * 100d);
+			var percent = (byte)Math.Round(((float)(cursor + amount) / size) * 100d);
 
 			_logger.Info($"{percent}% of the world sent...");
